@@ -1,5 +1,5 @@
 {
-  "name": "donation-clientsite",
+  "name": "client-site",
   "private": true,
   "version": "0.0.0",
   "type": "module",
@@ -10,38 +10,25 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@tanstack/react-query": "^5.7.0",
-    "axios": "^1.6.0",
-    "firebase": "^10.5.2",
-    "flowbite-react": "^0.6.4",
     "localforage": "^1.10.0",
-    "lottie-react": "^2.4.0",
     "match-sorter": "^6.3.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-helmet": "^6.1.0",
-    "react-helmet-async": "^1.3.0",
-    "react-hot-toast": "^2.4.1",
-    "react-icons": "^4.11.0",
-    "react-router-dom": "^6.18.0",
-    "react-table": "^7.8.0",
-    "react-toastify": "^9.1.3",
-    "reactive-button": "^1.3.15",
-    "sort-by": "^1.2.0",
-    "sweetalert2": "^11.9.0"
+    "react-router-dom": "^6.20.0",
+    "sort-by": "^1.2.0"
   },
   "devDependencies": {
-    "@types/react": "^18.2.15",
-    "@types/react-dom": "^18.2.7",
-    "@vitejs/plugin-react": "^4.0.3",
+    "@types/react": "^18.2.37",
+    "@types/react-dom": "^18.2.15",
+    "@vitejs/plugin-react": "^4.2.0",
     "autoprefixer": "^10.4.16",
-    "daisyui": "^3.9.4",
-    "eslint": "^8.45.0",
-    "eslint-plugin-react": "^7.32.2",
+    "daisyui": "^4.4.6",
+    "eslint": "^8.53.0",
+    "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.3",
+    "eslint-plugin-react-refresh": "^0.4.4",
     "postcss": "^8.4.31",
     "tailwindcss": "^3.3.5",
-    "vite": "^4.4.5"
+    "vite": "^5.0.0"
   }
 }
