@@ -1,4 +1,6 @@
-export const plugins = {
-  tailwindcss: {},
-  autoprefixer: {},
-};
+export default {
+  plugins: {
+    tailwindcss: {},
+    autoprefixer: {},
+  },
+}
